+++ packages/mcp/src/secrets/providers/dotenv/index.ts
@@ -1,0 +1,8 @@
+export { DotEnvProvider } from './provider.js';
+export type {
+  DotEnvProviderOptions,
+  DotEnvVariables,
+  DotEnvParserOptions,
+  LogicalLineParseResult,
+} from './types.js';
+export { parseDotEnvContent } from './parser.js';
